Explicitly reject unknown tool call parameters
