+++ servers/Fabric.Mcp.Server/changelog-entries/1786656227003.yaml
@@ -1,0 +1,3 @@
+changes:
+  - section: "Breaking Changes"
+    description: "Tool calls with unknown parameters are now explicitly rejected."
